pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentiment-optimizer"
version = "0.1.0"
description = "Real-Time Customer Sentiment & Support Optimizer"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gradio>=4.0.0",
    "confluent-kafka>=2.3.0",
    "anthropic>=0.25.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "rich>=13.7.0",
    "textblob>=0.17.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.hatch.build.targets.wheel]
packages = ["src"]